[refactor]: 테스트 코드의 접근 제한자 삭제
